Uses **Adapter pattern** to plugin different client backends in perf_analyzer harness.

~~For concurrency=1 and simple add_sub models(low latency models), these changes present a significant overhead.~~
~~**`throughput: 1943.5 infer/sec -> throughput: 1254 infer/sec`**~~
~~**Need to lessen the activity in the infer pipeline.**~~
 
~~For larger concurrency 8, the difference disappears as the requests masks the overhead. Same is true for larger models like inception_v2_graphdef where there is no noticeable difference for any concurrencies.~~

I was using an outdated devel container against newer build container for my experimentation. Somehow, the issue seems to have been resolved when using updated containers.

===============================================

**model_name** = graphdef_int32_int32_int32
**protocol** = grpc
**API** = synchronous
### Before
**Concurrency: 1, throughput: 1849.5 infer/sec, latency 539 usec**
### After
**Concurrency: 1, throughput: 1837 infer/sec, latency 543 usec**

===============================================
**model_name** = graphdef_int32_int32_int32
**protocol** = http
**API** = asynchronous
### Before
**Concurrency: 1, throughput: 2002 infer/sec, latency 481 usec**
### After
**Concurrency: 1, throughput: 2009.5 infer/sec, latency 480 usec**

===============================================

On an average I saw similar numbers from the two.